% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Modulon.Cores.LCMV}
\alias{Modulon.Cores.LCMV}
\title{Modulon.Cores.LCMV Data}
\format{
A character vector of 11 elements.
}
\source{
\url{https://github.com/icrespocasajus/ModulonCore}
}
\usage{
Modulon.Cores.LCMV
}
\description{
A dataset containing the connected components of each modulon.
}
\examples{
Modulon.Cores.LCMV
}
\keyword{datasets}
