% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonSat_functions.R
\name{plot.target.analysis.modulon}
\alias{plot.target.analysis.modulon}
\title{Plot results of the modulon target analysis}
\usage{
\method{plot}{target.analysis.modulon}(data, feature = "Redundancy", color = "YlGn")
}
\arguments{
\item{data}{List object with three elements ("Redundancy","Similarity" and "Overlap") with the corresponding matrices.}

\item{feature}{Target analysis feature to be displayed; one of c("Redundancy","Similarity","Overlap").}
}
\value{
A corrplot() object.
}
\description{
Generate the plot for the matrix with all pair-wise comparisons within a given modulon for one of the following features:"Redundancy","Similarity" and "Overlap".
}
\details{
This function generates a plot showing the results of the modulon target analysis of a given modulon.
}
\examples{
\dontrun{
if(interactive()){
results.target.analysis.modulon=target.analysis.modulon(net=network.TILs,mod=modulons.TILs,mod.query = '3')
plot.target.analysis.modulon(data=results.target.analysis.modulon,feature = 'Redundancy')
}
}
}
